+++ website/projects/scimaps/src/app/app.component.html
@@ -1,4 +1,4 @@
 <div class="news-item">
-  <sci-news-item></sci-news-item>
+  <sci-news-item [item]="newsItem"></sci-news-item>
 </div>
 <router-outlet></router-outlet>
