+++ website/projects/scimaps/src/app/app.component.scss
@@ -1,3 +1,3 @@
 .news-item {
-  width: 500px;
+  width: 31.25rem;
 }
