+++ website/projects/scimaps/src/app/app.component.ts
@@ -1,4 +1,6 @@
 import { Component } from '@angular/core';
+import { NewsItem } from './shared/news-item/news-item.model';
+
 
 @Component({
   selector: 'sci-root',
@@ -7,4 +9,13 @@ import { Component } from '@angular/core';
 })
 export class AppComponent {
   title = 'scimaps';
+
+  newsItem: NewsItem = {
+    title: 'The Places & Spaces: Mapping Science comes to Virginia Tech at the University Libraries',
+    date: 'MAR 2, 2020',
+    source: 'Library News',
+    institution: 'Virginia Tech',
+    thumbnail: '../../assets/images/rose.jpg',
+    pdfLink: 'link'
+  };
 }
