--- website/projects/scimaps/src/app/news-item/news-item.component.html
@@ -1,8 +1,0 @@
-<a class="news-item" [href]="item.pdfLink">
-  <img class="thumb" [src]="item.thumbnail" alt="image thumbnail">
-  <div class="info">
-    <div class="date">{{item.date}}</div>
-    <div class="title">{{item.title}}</div>
-    <div class="identifier">{{item.source}} at {{item.institution}}</div>
-  </div>
-</a>
