--- website/projects/scimaps/src/app/news-item/news-item.component.scss
@@ -1,41 +1,0 @@
-.news-item {
-  display: flex;
-  box-shadow: 0 1px 3px rgba(0,0,0,0.54);
-  text-decoration: none;
-  margin-top: 8px;
-  margin-bottom: 8px;
-  color: black;
-  border-radius: 2px;
-
-  .thumb {
-    width: 144px;
-    max-height: 120px;
-    border: 2px solid white;
-  }
-  
-  .info {
-    display: flex;
-    flex-direction: column;
-    font-size: 14px;
-    line-height: 24px;
-    padding: 16px;
-
-    .date {
-      font-size: 12px;
-      color: #4A5F68;
-    }
-
-    .title {
-      font-weight: 700;
-    }
-
-    .identifier {
-      font-weight: 300;
-      color: #4A5F68;
-    }
-  }
-}
-
-.news-item:hover {
-  box-shadow: 0 2px 6px rgba(0, 0, 0, 0.54)
-}
