+++ website/projects/scimaps/src/app/shared/news-item/news-item.model.ts
