+++ website/projects/scimaps/src/app/shared/news-item/news-item.theme.scss
@@ -1,0 +1,19 @@
+@use '~@angular/material/theming' as *;
+
+@mixin theme($theme) {
+  $primary: map-get($theme, primary);
+  $accent: map-get($theme, accent);
+
+  .sci-news-item {
+    .news-item {
+      color: black;
+    }
+    .identifier, .date {
+      color: mat-color($primary, darker);
+    }
+
+    .thumb {
+      border-color: white;
+    }
+  }
+}
